requirements.txt
streamlit
pandas
scikit-learn
PyPDF2
google-generativeai
python-dotenv